src/main/java/com/fastcampus/batch_campus/application/dormant/PreDormantItemProcessor.java
package com.fastcampus.batch_campus.application.dormant;

import com.fastcampus.batch_campus.batch.ItemProcessor;
import com.fastcampus.batch_campus.customer.Customer;
import org.springframework.stereotype.Component;

import java.time.LocalDate;

@Component
public class PreDormantItemProcessor implements ItemProcessor<Customer, Customer> {
    @Override
    public Customer process(Customer customer) {

        LocalDate targetDate = LocalDate.now()
                .minusDays(365)
                .plusDays(7);

        if (targetDate.equals(customer.getLoginAt().toLocalDate())) {
            return customer;
        } else {
            return null;
        }
    }
}
